requirements.txt
PyQt5
numpy
pillow
py2app
openai
sqlalchemy